requirements.txt
allure-pytest==2.9.43
allure-python-commons==2.9.43
attrs @ file:///tmp/build/80754af9/attrs_1620827162558/work
baidu-api==0.0.2
brotlipy==0.7.0
certifi==2020.6.20
cffi @ file:///Users/runner/miniforge3/conda-bld/cffi_1613414076490/work
chardet @ file:///Users/runner/miniforge3/conda-bld/chardet_1610093454189/work
cryptography @ file:///Users/runner/miniforge3/conda-bld/cryptography_1616851494399/work
et-xmlfile==1.0.1
execnet==1.9.0
Faker @ file:///tmp/build/80754af9/faker_1624291529673/work
idna @ file:///home/linux1/recipes/ci/idna_1610986105248/work
importlib-metadata @ file:///tmp/build/80754af9/importlib-metadata_1602276842396/work
iniconfig @ file:///home/linux1/recipes/ci/iniconfig_1610983019677/work
install==1.3.4
jdcal==1.4.1
jmespath==0.10.0
more-itertools @ file:///tmp/build/80754af9/more-itertools_1622818384463/work
Naked==0.1.31
numpy @ file:///Users/runner/miniforge3/conda-bld/numpy_1624399904045/work
olefile==0.46
openpyxl @ file:///tmp/build/80754af9/openpyxl_1615411699337/work
packaging @ file:///tmp/build/80754af9/packaging_1611952188834/work
pandas==1.2.5
Pillow @ file:///Users/runner/miniforge3/conda-bld/pillow_1621288272555/work
pluggy==0.12.0
py @ file:///tmp/build/80754af9/py_1607971587848/work
pycparser @ file:///tmp/build/80754af9/pycparser_1594388511720/work
pycryptodome==3.10.1
PyMySQL @ file:///home/conda/feedstock_root/build_artifacts/pymysql_1610208984992/work
pyOpenSSL @ file:///tmp/build/80754af9/pyopenssl_1608057966937/work
pyparsing @ file:///home/linux1/recipes/ci/pyparsing_1610983426697/work
PySocks @ file:///Users/runner/miniforge3/conda-bld/pysocks_1610291456365/work
pytesseract==0.3.7
pytest==6.2.4
pytest-assume==2.4.3
pytest-forked==1.3.0
pytest-html==3.1.1
pytest-metadata==1.11.0
pytest-repeat==0.9.1
pytest-rerunfailures==10.1
pytest-xdist==2.3.0
python-dateutil @ file:///home/ktietz/src/ci/python-dateutil_1611928101742/work
pytz @ file:///tmp/build/80754af9/pytz_1612215392582/work
PyYAML==5.4.1
redis==3.5.3
requests @ file:///tmp/build/80754af9/requests_1608241421344/work
selenium==3.141.0
shellescape==3.8.1
six @ file:///tmp/build/80754af9/six_1623709665295/work
text-unidecode==1.3
toml @ file:///tmp/build/80754af9/toml_1616166611790/work
urllib3==1.26.5
xlrd @ file:///home/conda/feedstock_root/build_artifacts/xlrd_1610224409810/work
xlwt==1.3.0
zipp @ file:///tmp/build/80754af9/zipp_1615904174917/work
zmail==0.2.8
